requirements.txt
torch~=1.10.2
Pillow~=9.0.1
torchvision~=0.12.0
matplotlib~=3.5.1
numpy~=1.22.2
tqdm~=4.62.3
scipy~=1.8.0